Extract Rugged code into a Changeset adapter
Move all direct Rugged usage out of Changeset into a new
RuggedAdapter. Changeset now delegates workdir lookup, file/line
diff iteration, and emptiness checks to an adapter, keeping
filtering and validation logic in Changeset itself. No behavior
change.
